requirements.txt
fastapi>=0.109.0
uvicorn[standard]==0.24.0
google-genai==1.27.0
duckduckgo-search==4.1.1
cachetools==5.3.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
httpx
python-dotenv==1.0.0
